apps/requirements.txt
backcall==0.1.0
wheel
certifi==2019.9.11
Click==7.0
colorama==0.4.1
cycler==0.10.0
dash==1.20.0
dash-bootstrap-components==0.12.1a4
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
decorator==4.4.2
Django==2.2.5
Faker==2.0.1
Flask==1.1.2
Flask-Compress==1.5.0
flaskapp==0.0.9
future==0.17.1
image==1.5.27
ipython==7.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
JayDeBeApi==1.1.1
jedi==0.15.1
Jinja2==2.10.1
joblib==0.13.2
jpype1==0.7.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.17.1
pandas==0.25.1
parso==0.5.1
pathlib==1.0.1
pickleshare==0.7.5
Pillow==6.1.0
pipenv==2018.11.26
plotly==4.7.1
prompt-toolkit==2.0.9
psycopg2==2.8.3
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
retrying==1.3.3
Rtree==0.8.3
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
text-unidecode==1.2
traitlets==4.3.2
virtualenv==16.7.5
virtualenv-clone==0.5.3
wcwidth==0.1.7
Werkzeug==0.15.6
gunicorn==20.0.4
folium==0.11.0
cufflinks==0.17.3
matplotlib_venn
unidecode
wordcloud
flask_mail
sql
mysqlclient
mysql-connector-python
pymysql
cryptography